change logic & add loading animation

+++ src/ps_camp/sql_models/bank_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, Enum
+from sqlalchemy import Column, BigInteger, String, ForeignKey, DateTime, Enum
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 import enum
@@ -29,7 +29,7 @@ class BankAccount(Base):
     owner_id = Column(String, nullable=False)
     owner_type = Column(Enum(OwnerType), nullable=False)
     account_number = Column(String, unique=True, nullable=False)
-    balance = Column(Integer, nullable=False, default=0)
+    balance = Column(BigInteger, nullable=False, default=0)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
 
     outgoing_transactions = relationship("Transaction", back_populates="from_account", foreign_keys="Transaction.from_account_id")
@@ -42,7 +42,7 @@ class Transaction(Base):
     id = Column(String, primary_key=True)
     from_account_id = Column(String, ForeignKey("bank_accounts.id"))
     to_account_id = Column(String, ForeignKey("bank_accounts.id"))
-    amount = Column(Integer, nullable=False)
+    amount = Column(BigInteger, nullable=False)
     note = Column(String, nullable=True)
     transaction_type = Column(Enum(TransactionType), nullable=False)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
